Refactor Windows drag and drop implementation
